requirements.txt
# PhishingDetector-AI Requirements
# Author: Ayi NEDJIMI

python>=3.11
torch>=2.0.0
transformers>=4.35.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
requests>=2.31.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
structlog>=23.2.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
